requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.20.0
streamlit>=1.32.0
feedparser>=6.0.0
python-dateutil>=2.9.0
